Fix a regression on forgotten delayed read of a type info.
There was a regression on gdb.base/type-opaque.exp.

Currently the delayed reading is determined according to `.debug_aranges'.  But
in this case Archer needs to read-in type information - it has no associated
address range.  In the case of this testcase it could be fixed by reading in
also the indexing section `.debug_pubnames'.  Still it does not solve `static'
names which are not contained in `.debug_pubnames' - GCC should produce also
(currently not present in the DWARF standard) `.debug_privnames' suggested by
Tom Tromey on <archer@sourceware.org> before for this purpose.

Assuming (did not measure it so far) this patch will nullify any performance
effect of this archer-tromey-delayed-symfile branch.

	* symtab.c (basic_lookup_transparent_type): Call ALL_PSYMTABS_REQUIRED
	instead of ALL_PSYMTABS.  New comment.