This patch brings no new functionality.  It could be enabled only during some --enable-debug builds but there is currently no such configure option.
Found it caught some bugs during the development (by trying to double-free some
types).  It can be also used to verify there are no types leaks.

There is now a fix to free anything at all.  When the function was written
there was an older implementation of type_decref which was incorrectly freeing
the zero-reference-count types immediately (and not just on the garbage
collection by free_all_types).

gdb/
2009-04-13  Jan Kratochvil  <jan.kratochvil@redhat.com>

	* value.c (value_history_cleanup): New function.
	(_initialize_values): Setup a value_history_cleanup final call.