Fix compilation warning.
gcc-4.4.2-20.fc12.x86_64 started to error out on:
gdbtypes.c: In function ‘delete_main_type’:
gdbtypes.c:3149: error: the address of ‘cplus_struct_default’ will always evaluate as ‘true’

It would be possible to fix it by:
-  gdb_assert (!HAVE_CPLUS_STRUCT (type) || !TYPE_CPLUS_SPECIFIC (type));
+  gdb_assert (!HAVE_CPLUS_STRUCT (type) || TYPE_CPLUS_SPECIFIC (type) == NULL);

But at the same due since the introduction of TYPE_SPECIFIC_CPLUS_STUFF by
	Re: [RFA/DWARF] Add DW_AT_GNAT_descriptive_type support
	http://sourceware.org/ml/gdb-patches/2010-01/msg00243.html

TYPE_CPLUS_SPECIFIC can no longer ever be NULL and therefore this conditional
can be removed.

Another point is that there should be now rather:
  gdb_assert (TYPE_SPECIFIC_FIELD (type) == TYPE_SPECIFIC_NONE)
as none of the extensions is supported but this garbage collecting code is
currently not in use anyway.

	* gdbtypes.c (delete_main_type): Remove !TYPE_CPLUS_SPECIFIC assertion.