Remove some FIXME comments
This removes some outdated FIXME comments.

The original design seems promising:

    http://sourceware.org/ml/gdb-patches/2003-11/msg00154.html

... though not fully described.  Maybe the intent was to implement it
by adding the index->ops mapping to the symtab.  Or, maybe a global
table was intended, with different symbol readers using different
parts of the index space.

However, since that time, SYMBOL_REGISTER_OPS was added.  This seems
to invalidate the earlier design:

    http://sourceware.org/ml/gdb-patches/2009-05/msg00420.html

It does seem to me that 'address_class' is a moderately confused
concept; or at least that LOC_COMPUTED is.

Nevertheless, I think that since only negative progress has been made
on this front in the last 8 years, it is best to zap the comments and
continue on with what we have.  If someone wants to revamp this area,
the comments are not likely to help anyhow.

	* ax-gdb.c (gen_var_ref): Remove comment.
	* findvar.c (symbol_read_needs_frame): Remove comment.
	(default_read_var_value): Likewise.
	* printcmd.c (address_info): Remove comment.