record-btrace: extend unwinder
Extend the always failing unwinder to provide the PC based on the call structure
detected in the branch trace.

There are several open points:

An assertion in get_frame_id at frame.c:340 requires that a frame provides a
stack address.  The record-btrace unwinder can't provide this since the trace
does not contain data.  I incorrectly set stack_addr_p to 1 to avoid the
assertion.

When evaluating arguments for printing the stack back trace, there's an ugly
error displayed: "error reading variable: can't compute CFA for this frame".
The error is correct, we can't compute the CFA since we don't have the stack at
that time, but it is rather annoying at this place and makes the back trace
difficult to read.

Now that we set the PC to a different value and provide a fake unwinder, we have
the potential to affect almost every other command.  How can this be tested
sufficiently?  I added a few tests for the intended functionality, but nothing
so far to ensure that it does not break some other command when used in this
context.

CC: Eli Zaretskii  <eliz@gnu.org>
2013-04-24  Markus Metzger  <markus.t.metzger@intel.com>

	* frame.h (enum frame_type) <BTRACE_FRAME>: New.
	* record-btrace.c: Include hashtab.h.
	(btrace_get_bfun_name): New.
	(btrace_call_history): Call btrace_get_bfun_name.
	(enum btrace_frame_flag): New.
	(struct btrace_frame_cache): New.
	(bfcache): New.
	(bfcache_hash, bfcache_eq, bfcache_new): New.
	(btrace_get_frame_function): New.
	(record_btrace_frame_unwind_stop_reason): Allow unwinding.
	(record_btrace_frame_this_id): Compute own id.
	(record_btrace_frame_prev_register): Provide PC, throw_error
	for all other registers.
	(record_btrace_frame_sniffer): Detect btrace frames.
	(record_btrace_frame_dealloc_cache): New.
	(record_btrace_frame_unwind): Add new functions.
	(_initialize_record_btrace): Allocate cache.
	* btrace.c (btrace_clear): Call reinit_frame_cache.
	* NEWS: Announce it.

testsuite/
	* gdb.btrace/record_goto.exp: Add backtrace test.
	* gdb.btrace/tailcall.exp: Add backtrace test.

Change-Id: Ic762c82c48baa77bc94755b808896a6e7f174737