introduce scoped cleanups
This adds scoped cleanups to gdb.  A scoped cleanup is a more
efficient, and checkable, alternative to the common idiom:

    cleanup = make_cleanup (null_cleanup, NULL);
    ...
    do_cleanups (cleanup);

... where the cleanup is always run when exiting the scope.

The efficiency comes from stack allocating the scoped cleanup.  This
is probably minor.  However, I've noticed myself sometimes avoiding
null cleanups on this basis, and it is nice to erase that bias.

The checking comes from using a GCC extension when available.  This
check ensures that the scoped cleanup was either run or discarded when
the scope exits.