add target method delegation
This patch replaces some code in the record targets with target method
delegation.

Right now there are two latent problems in the record target.

First, record-full.c stores pointers to many target methods when the
record target is pushed.  Then it later delegates some calls via
these.  This is wrong because it violates the target stack contract.
In particular it is ok to unpush a target at any stratum, but
record-full does not keep track of this, so it could potentially call
into an unpushed target.

Second, RECORD_IS_USED and some other spots look at
current_target.to_stratum to determine whether a record target is in
use.  This is bad because arch_stratum is greater than record_stratum.

To fix the first problem, this patch introduces a handful of
target_delegate_* functions, which forward calls further down the
target stack.

To fix the second problem, this patch adds find_target_at to determine
whether a target appears at a given stratum.  This may seem like
overkill somehow, but I have a subsequent patch series (see archer.git
tromey/multi-target) that uses it more heavily.

	* record-full.c (record_full_beneath_to_resume_ops)
	(record_full_beneath_to_resume, record_full_beneath_to_wait_ops)
	(record_full_beneath_to_wait)
	(record_full_beneath_to_store_registers_ops)
	(record_full_beneath_to_store_registers)
	(record_full_beneath_to_xfer_partial_ops)
	(record_full_beneath_to_xfer_partial)
	(record_full_beneath_to_insert_breakpoint)
	(record_full_beneath_to_remove_breakpoint)
	(record_full_beneath_to_stopped_by_watchpoint)
	(record_full_beneath_to_stopped_data_address)
	(record_full_beneath_to_async, tmp_to_resume_ops, tmp_to_resume)
	(tmp_to_wait_ops, tmp_to_wait, tmp_to_store_registers_ops)
	(tmp_to_store_registers, tmp_to_xfer_partial_ops)
	(tmp_to_xfer_partial, tmp_to_insert_breakpoint)
	(tmp_to_remove_breakpoint, tmp_to_stopped_by_watchpoint)
	(tmp_to_stopped_data_address, tmp_to_async): Remove.
	(record_full_open_1, record_full_open): Update.  Use RECORD_IS_USED.
	(record_full_resume, record_full_wait_1)
	(record_full_stopped_by_watchpoint, record_full_stopped_data_address)
	(record_full_store_registers, record_full_xfer_partial)
	(record_full_insert_breakpoint, record_full_remove_breakpoint)
	(record_full_async, record_full_can_async_p, record_full_is_async_p)
	(record_full_core_xfer_partial): Use target delegation.
	* record.c (find_record_target): Use find_target_at.
	* record.h (RECORD_IS_USED): Use find_target_at.
	* target.c (update_current_target): Use target_delegate_xfer_partial.
	(target_delegate_xfer_partial): Now public.  Renamed from...
	(current_xfer_partial): ...this.  Remove.
	(target_delegate_async, target_delegate_is_async_p)
	(target_delegate_can_async_p, target_delegate_insert_breakpoint)
	(target_delegate_remove_breakpoint, target_delegate_wait)
	(target_delegate_resume, find_target_at)
	(target_delegate_store_registers)
	(target_delegate_stopped_by_watchpoint)
	(target_delegate_stopped_data_address): New functions.
	* target.h (target_delegate_async, target_delegate_is_async_p)
	(target_delegate_can_async_p, target_delegate_insert_breakpoint)
	(target_delegate_remove_breakpoint, target_delegate_wait)
	(target_delegate_resume, find_target_at)
	(target_delegate_store_registers)
	(target_delegate_stopped_by_watchpoint)
	(target_delegate_stopped_data_address)
	(target_delegate_xfer_partial): Declare.